+++ src/inputs.ts
@@ -1,7 +1,7 @@
 import * as core from "@actions/core";
 
 
-const INPUT_NAME_COMMENTS = "comments";
+const INPUT_NAME_COMMENT = "comment";
 const INPUT_NAME_CONFIG_PATH = "configuration-path";
 const INPUT_NAME_CONVENTIONAL_COMMITS = "conventional-commits";
 const INPUT_NAME_DRY_RUN = "dry-run";
@@ -18,17 +18,17 @@ const STRING = "string";
 const TRUE = "true";
 
 const CONVENTIONAL_COMMIT_TITLE = "chore: optimize {{optimizedCount}} SVG(s)";
-const DEFAULT_COMMIT_DESCRIPTION = "Optimized SVG(s):\n{{filesList}}";
+const DEFAULT_COMMIT_BODY = "Optimized SVG(s):\n{{filesList}}";
 const DEFAULT_COMMIT_TITLE = "Optimize {{optimizedCount}} SVG(s) with SVGO";
 const DEFAULT_COMMENT = "SVG(s) automatically optimized using [SVGO](https://github.com/svg/svgo) :sparkles:\n\n{{filesTable}}";
 
 
 export type RawActionConfig = {
-  readonly comments?: boolean | string;
+  readonly comment?: boolean | string;
   readonly commit?: {
     readonly conventional?: boolean;
     readonly title?: string;
-    readonly description?: string;
+    readonly body?: string;
   };
   readonly "dry-run"?: boolean;
   readonly ignore?: string;
@@ -47,7 +47,7 @@ export function getRepoToken(): string {
 export class ActionConfig {
 
   public readonly comment: string;
-  public readonly commitDescription: string;
+  public readonly commitBody: string;
   public readonly commitTitle: string;
   public readonly enableComments: boolean;
   public readonly ignoreGlob: string;
@@ -56,7 +56,7 @@ export class ActionConfig {
 
   constructor(config: RawActionConfig = { }) {
     this.comment = ActionConfig.getCommentValue(config);
-    this.commitDescription = ActionConfig.getCommitDescription(config);
+    this.commitBody = ActionConfig.getCommitBody(config);
     this.commitTitle = ActionConfig.getCommitTitle(config);
     this.enableComments = ActionConfig.getEnableCommentsValue(config);
     this.ignoreGlob = ActionConfig.getIgnoreGlob(config);
@@ -65,8 +65,8 @@ export class ActionConfig {
   }
 
   private static getCommentValue(config: RawActionConfig): string {
-    const value = (config.comments !== undefined) ?
-      config.comments : core.getInput(INPUT_NAME_COMMENTS, NOT_REQUIRED);
+    const value = (config.comment !== undefined) ?
+      config.comment : core.getInput(INPUT_NAME_COMMENT, NOT_REQUIRED);
     if (typeof value === STRING && value !== TRUE) {
       // If the value is (the string) `"false"` comments will be disabled, so it
       // does not matter that the comment template is `"false"`. If the value is
@@ -78,9 +78,9 @@ export class ActionConfig {
     }
   }
 
-  private static getCommitDescription(config: RawActionConfig): string {
-    return (config.commit?.description !== undefined) ?
-      config.commit.description : DEFAULT_COMMIT_DESCRIPTION;
+  private static getCommitBody(config: RawActionConfig): string {
+    return (config.commit?.body !== undefined) ?
+      config.commit.body : DEFAULT_COMMIT_BODY;
   }
 
   private static getCommitTitle(config: RawActionConfig): string {
@@ -98,7 +98,7 @@ export class ActionConfig {
   }
 
   private static getEnableCommentsValue(config: RawActionConfig): boolean {
-    return this.normalizeBoolOption(config.comments as boolean, INPUT_NAME_COMMENTS, true);
+    return this.normalizeBoolOption(config.comment as boolean, INPUT_NAME_COMMENT, true);
   }
 
   private static getIgnoreGlob(config: RawActionConfig): string {