+++ test/mocks/inputs.mock.ts
@@ -1,6 +1,6 @@
 export const ActionConfig = jest.fn()
   .mockImplementation(() => ({
-    commitDescription: "Optimized SVGs:\n{{fileList}}",
+    commitBody: "Optimized SVGs:\n{{fileList}}",
     commitTitle: "Optimize {{optimizedCount}} SVG(s) with SVGO",
     ignoreGlob: "",
     isDryRun: false,