+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Versioning].
 
 ## [Unreleased]
 
+- (!) Rename the `comments` option to `comment`. ([#190])
+- (!) Rename the `commit.description` options to `commit.body`. ([#190])
 - Add `ignoredCount` as templating variable. ([#191])
 
 ## [0.4.4] - 2020-04-12
@@ -14,7 +16,7 @@ Versioning].
 - Add functionality to ignore SVGs based on a glob. ([#175], [#179])
 - Add option to customize Pull Request comments. ([#184], [#186])
 - Update the documentation for the Action's options. ([#185])
-- Add the files table template value for commit descriptions. ([#187])
+- Add the files table template value for commit message bodies. ([#187])
 
 ## [0.4.3] - 2020-04-10
 
@@ -37,7 +39,7 @@ Versioning].
 ## [0.3.2] - 2020-03-13
 
 - Re-enable the Action from a Pull Request comment. ([#129])
-- Fix default commit message description. ([#130])
+- Fix the default commit message body. ([#130])
 
 ## [0.3.1] - 2020-03-12
 
@@ -125,4 +127,5 @@ Versioning].
 [#185]: https://github.com/ericcornelissen/svgo-action/pull/185
 [#186]: https://github.com/ericcornelissen/svgo-action/pull/186
 [#187]: https://github.com/ericcornelissen/svgo-action/pull/187
+[#190]: https://github.com/ericcornelissen/svgo-action/pull/190
 [#191]: https://github.com/ericcornelissen/svgo-action/pull/191